services/notification-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
pymongo==4.5.0
celery==5.3.4
redis==5.0.1
jinja2==3.1.2
httpx==0.24.1
python-jose==3.3.0
pyjwt==2.8.0
python-multipart==0.0.6
email-validator==2.0.0
aiosmtplib==2.0.2
python-telegram-bot==20.6
python-dotenv==1.0.0